backend/bff_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
